RENAME: checksum_validator -> _checksum_validator.py
